Change the return type of cfile_datastream::tellp() to size_t

+++ include/fc/io/cfile.hpp
@@ -145,8 +145,8 @@ public:
 
    bool get( char& c ) { return read(&c, 1); }
 
-   long tellp() const {
-      return cf.tellp();
+   size_t tellp() const {
+      return static_cast<size_t>(cf.tellp());
    }
 
 private:
